Time: 22 ms (49.07%), Space: 46.1 MB (12.41%) - LeetHub

+++ 0518-coin-change-ii/0518-coin-change-ii.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
 
-   long change(int amount, vector<int>& coins) {
+   int change(int amount, vector<int>& coins) {
         int n=coins.size();
         vector<vector<double>>dp(n,vector<double>(amount+1,0));
         for(int i=0;i<=amount;i++){
@@ -17,6 +17,6 @@ public:
                 dp[i][j]=pick+notpick;
             }
         }
-        return (dp[n-1][amount]);
+        return int(dp[n-1][amount]);
     }
 };
